+++ epictrack-web/src/components/workPlan/status/StatusContainer/StatusNotes.tsx
@@ -1,0 +1,16 @@
+import RichTextEditor from "../../../shared/richTextEditor";
+
+const StatusNotes = () => {
+  const handleNotesChange = (value: string) => {
+    return null;
+  };
+
+  return (
+    <RichTextEditor
+      handleEditorStateChange={handleNotesChange}
+      // initialRawEditorState={}
+    />
+  );
+};
+
+export default StatusNotes;
