+++ epictrack-web/src/components/workPlan/status/StatusContext.tsx
@@ -1,0 +1,18 @@
+import { createContext, useContext } from "react";
+import { Status } from "../../../models/status";
+import React from "react";
+import { WorkplanContext } from "../WorkPlanContext";
+
+// interface StatusContextProps {}
+
+// export const StatusContext = createContext<StatusContextProps>({});
+
+// export const StatusProvider = ({
+//   children,
+// }: {
+//   children: JSX.Element | JSX.Element[];
+// }) => {
+//   return (
+//     <StatusContext.Provider value={null}>{children}</StatusContext.Provider>
+//   );
+// };
