+++ epictrack-web/src/components/workPlan/status/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import StatusContainer from "./StatusContainer";
+
+const Status = () => {
+  return <StatusContainer />;
+};
+
+export default Status;
